requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
django-filter~=24.3
google-api-python-client~=2.149.0
google-auth-httplib2
google-auth-oauthlib